dev::Create XML Preview Tool Window and Identify Data Source
